Preparing for LT 1.0.0

+++ src/app/location/location/location.component.ts
@@ -1,9 +1,12 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { Location, Light, Plant } from 'src/app/interfaces';
+import { Location, Light } from 'src/app/interfaces';
 import { BreadcrumbService } from 'src/app/breadcrumb/breadcrumb.service';
 import { ApiService } from 'src/app/shared/api/api.service';
 import { AuthService } from 'src/app/auth/auth.service';
+import { catchError, EMPTY, map } from 'rxjs';
+import { HttpErrorResponse } from '@angular/common/http';
+import { ErrorHandlerService } from 'src/app/error-handler/error-handler.service';
 
 @Component({
   selector: 'location',
@@ -11,8 +14,7 @@ import { AuthService } from 'src/app/auth/auth.service';
   styleUrls: ['./location.component.scss']
 })
 export class LocationComponent implements OnInit {
-  private id!: number;
-  isValidId!: boolean;
+  private id?: number;
   location!: Location;
   confirmDelete: boolean = false;
   owned: boolean = false;
@@ -22,7 +24,8 @@ export class LocationComponent implements OnInit {
     private router: Router,
     private api: ApiService,
     private auth: AuthService,
-    private breadcrumb: BreadcrumbService
+    private breadcrumb: BreadcrumbService,
+    private errorHandler: ErrorHandlerService
   ) { }
 
   ngOnInit(): void {
@@ -30,38 +33,43 @@ export class LocationComponent implements OnInit {
     this.id = paramId ? +paramId : NaN;
 
     if (this.id) {
-      this.isValidId = true;
 
-      this.api.getLocation(this.id, true).subscribe({
-        next: (location) => {
+      this.api.getLocation(this.id, true).pipe(
+        map((location: Location) => {
           this.location = location;
           this.breadcrumb.setNavigation([
             { id: 'location', name: this.location.name, link: ['/location', this.id] },
           ])
 
           this.owned = (this.auth.user$.getValue()?.id === this.location.ownerId) ? true : false;
-        },
-        error: (error) => {
-          if (error.msg === 'LOCATION_NOT_FOUND') this.isValidId = false;
-        }
-      });
+        }),
+        catchError((err: HttpErrorResponse) => {
+          if (err.error?.msg === 'LOCATION_NOT_FOUND') this.errorHandler.push($localize `:@@location.invalid:Location invalid or not found.`);
+          else this.errorHandler.push($localize `:@@errors.server:Server error`);
+          
+          this.router.navigateByUrl('/');
+  
+          return EMPTY;
+        })
+      ).subscribe();
     }
-    else this.isValidId = false;
   }
 
   edit(): void {
-    this.router.navigate(['/location', 'edit', this.id]);
+    if (this.id) this.router.navigate(['/location', 'edit', this.id]);
   }
 
   delete(): void {
-    this.api.deleteLocation(this.id).subscribe({
-      next: (data) => {
-        if (data.msg === 'LOCATION_REMOVED') this.router.navigate(['/']);
-      },
-      error: (err) => {
-        if (err.msg === 'LOCATION_NOT_VALID') console.log('error');
-      }
-    })
+    if (this.id) {
+      this.api.deleteLocation(this.id).subscribe({
+        next: (res) => {
+          if (res.msg === 'LOCATION_REMOVED') this.router.navigate(['/']);
+        },
+        error: (err) => {
+          if (err.msg === 'LOCATION_NOT_VALID') this.errorHandler.push($localize `:@@location.invalid:Location invalid or not found`)
+        }
+      })
+    }
   }
 
   getLightName(): string {
