+++ src/app/plant/plant.service.ts
@@ -1,6 +1,6 @@
 import { HttpErrorResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { BehaviorSubject, catchError, map, Observable, throwError } from 'rxjs';
+import { BehaviorSubject, catchError, EMPTY, map, Observable, throwError } from 'rxjs';
 import { AuthService } from '../auth/auth.service';
 import { Plant } from '../interfaces';
 import { ApiService } from '../shared/api/api.service';
@@ -9,20 +9,19 @@ import { ApiService } from '../shared/api/api.service';
   providedIn: 'root'
 })
 export class PlantService {
-  plant$: BehaviorSubject<Plant> = new BehaviorSubject<Plant>({} as Plant);
+  plant$: BehaviorSubject<Plant | null> = new BehaviorSubject<Plant | null>(null);
   owned: boolean = false;
 
   constructor(
     private api: ApiService,
     private auth: AuthService
-    ) { }
+  ) { }
 
   get(id: number): Observable<any> {
     return this.api.getPlant(id).pipe(
       map((plant: Plant) => {
         this.owned = (this.auth.user$.getValue()?.id === plant.ownerId);
 
-        // FIXME: do it in ApiService?
         if (plant.customName) plant.visibleName = plant.customName;
         else if (plant.specie?.name) plant.visibleName = plant.specie.name;
         else plant.visibleName = $localize `:@@general.unnamedPlant:Unnamed plant ${plant.id}:plantId:`;
@@ -32,51 +31,63 @@ export class PlantService {
         return plant;
       }),
       catchError((HttpError: HttpErrorResponse) => {
-        this.plant$.next({} as Plant);
+        this.plant$.next(null);
 
-        return throwError(() => HttpError.error);
+        return throwError(() => HttpError);
       })
     );
   }
 
-  update(plant: Plant): Observable<any> {
-    return this.api.updatePlant(plant).pipe(
+  update(plant: Plant, options?: any): Observable<any> {
+    return this.api.updatePlant(plant, options).pipe(
       map((plant: Plant) => {
         const current = this.plant$.getValue();
-        plant.photos = current.photos;
-        this.plant$.next(plant);
+
+        if (current) {
+          plant.photos = current.photos;
+          this.plant$.next(plant);
+        }
       }),
       catchError((HttpError: HttpErrorResponse) => {
-        this.plant$.next({} as Plant);
+        this.plant$.next(null);
 
-        return throwError(() => HttpError.error);
+        return throwError(() => HttpError);
       })
     );
   }
 
   delete(): Observable<any> {
-    const { id } = this.plant$.getValue();
+    const id = this.plant$.getValue()?.id;
 
-    return this.api.deletePlant(id);
+    if (id) return this.api.deletePlant(id);
+    else return EMPTY;
   }
 
   fertilize(): Observable<any> {
-    const { id } = this.plant$.getValue();
-    const updatedPlant = {
-      id: id,
-      fertLast: new Date()
-    } as Plant;
+    const id = this.plant$.getValue()?.id;
 
-    return this.update(updatedPlant);
+    if (id) {
+      const updatedPlant = {
+        id: id,
+        fertLast: new Date()
+      } as Plant;
+      return this.update(updatedPlant);
+    }
+
+    return EMPTY;
   }
 
   water(): Observable<any> {
-    const { id } = this.plant$.getValue();
-    const updatedPlant = {
-      id: id,
-      waterLast: new Date()
-    } as Plant;
+    const id = this.plant$.getValue()?.id;
 
-    return this.update(updatedPlant);
+    if (id) {
+      const updatedPlant = {
+        id: id,
+        waterLast: new Date()
+      } as Plant;
+      return this.update(updatedPlant);
+    }
+
+    return EMPTY;
   }
 }
